Add file via upload
This script takes a defined URL and puts all the urls within it into a urls array. Then, the script goes through the urls and scrapes specific data to put it into a csv. It was written to easily see specific data for 150+ audio program web pages in a spreadsheet.